cert: move cert_to_ssh_key_send/recv() to ssh responder
The cert_to_ssh_key_send/recv() request is currently only used by the
SSD responder and adds an unneeded dependency to libsss_certmap to
lib_sss_cert.

Resolves: https://github.com/SSSD/sssd/issues/1030
