+++ src/util/cert.h
@@ -54,16 +54,4 @@ errno_t get_ssh_key_from_cert(TALLOC_CTX *mem_ctx,
 
 errno_t get_ssh_key_from_derb64(TALLOC_CTX *mem_ctx, const char *derb64,
                                 uint8_t **key_blob, size_t *key_size);
-
-struct tevent_req *cert_to_ssh_key_send(TALLOC_CTX *mem_ctx,
-                                        struct tevent_context *ev,
-                                        const char *logfile, time_t timeout,
-                                        const char *ca_db,
-                                        struct sss_certmap_ctx *sss_certmap_ctx,
-                                        size_t cert_count,
-                                        struct ldb_val *bin_certs,
-                                        const char *verify_opts);
-
-errno_t cert_to_ssh_key_recv(struct tevent_req *req, TALLOC_CTX *mem_ctx,
-                             struct ldb_val **keys, size_t *valid_keys);
 #endif /* __CERT_H__ */
